Some fixes and refactoring
